size

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long w;
 	float f;
 
-	printf("Size of char: %zu byte(s)\n", Sizeof(v));
-	printf("Size of an int: %zu byte(S)\n", Sizeof(i));
-	printf("Size of a long int: %zu byte(s)\n", Sizeof(z));
-	printf("Size of a long long int: %zu byte(s)\n", Sizeof(w));
-	printf("Size of a float: %zu byte(s)\n", Sizeof(f));
+	printf("Size of char: %zu byte(s)\n", (unsigned long)Sizeof(v));
+	printf("Size of an int: %zu byte(S)\n", (unsigned long)Sizeof(i));
+	printf("Size of a long int: %zu byte(s)\n", (unsigned long)Sizeof(z));
+	printf("Size of a long long int: %zu byte(s)\n", (unsigned long)Sizeof(w));
+	printf("Size of a float: %zu byte(s)\n", (unsigned long) Sizeof(f));
 	return (0);
 }
